rename com.dwolla.sbt.awslambda to com.dwolla.sbt.s3
